workers/requirements.txt
openai-whisper>=20240930
anthropic>=0.52.0
numpy>=1.26
